+++ src/main/java/com/gomo/app/common/domain/service/OrderChangeable.java
@@ -1,4 +1,4 @@
-package com.gomo.app.common.domain;
+package com.gomo.app.common.domain.service;
 
 public interface OrderChangeable {
